refactor: moved converted files into  subfolder
