+++ src/main/kotlin/io/github/agentseek/common/Circle2d.kt
