+++ src/main/kotlin/io/github/agentseek/common/GameMath.kt
