+++ src/main/kotlin/io/github/agentseek/common/Point2d.kt
