+++ src/main/kotlin/io/github/agentseek/common/Shape2d.kt
